fix ArrayIndexOutOfBoundsException thrown while parsing devhelp2 #41
- add new SAXParser to handle the new gtk-doc devhelp2 format where
  references are to a single file and do not include links
- rename HTMLSaxParser to HTMLSAXParserOld and move to new file
- create new file HTMLSAXParser which parses the new format
- support for devhelp2 files only, if no links exist for devhelp
  file, bail on parsing
