+++ libhover/org.eclipse.linuxtools.cdt.libhover.devhelp/src/org/eclipse/linuxtools/internal/cdt/libhover/devhelp/DevHelpSAXParser.java
@@ -1,0 +1,21 @@
+/*******************************************************************************
+ * Copyright (c) 2022 Red Hat Inc. and others.
+ * 
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ * Red Hat Inc. - Initial implementation
+ *******************************************************************************/
+package org.eclipse.linuxtools.internal.cdt.libhover.devhelp;
+
+import java.util.TreeMap;
+
+import org.eclipse.linuxtools.cdt.libhover.FunctionInfo;
+
+interface DevHelpSAXParser {
+    public TreeMap<String, FunctionInfo> getFunctionInfos();
+}
